small bug fix
git-svn-id: https://publicdomain.svn.codeplex.com/svn@32402 11aa8812-6de0-41ed-92bd-ec94540cc92c

+++ PublicDomain/Config/ConfigurationValues.cs
@@ -343,7 +343,7 @@ namespace PublicDomain.Config
         /// <param name="key">The key.</param>
         /// <param name="defaultValue">The default value.</param>
         /// <returns></returns>
-        public double GetDouble(string key, int defaultValue)
+        public double GetDouble(string key, double defaultValue)
         {
             string val = this[key];
             return val == null ? defaultValue : double.Parse(val);
